// src/types/core.ts
export interface DeviceDetails {
    sensor_device_id: number;
    sensor_device_name: string;
    camera_device_id: number;
    camera_device_name: string;
}

export interface LocationDetails {
    location_id: number;
    location_name: string;
}

export interface SensorDeviceDetails {
    sensor_device_id: number;
    sensor_device_name: string;
}

export interface CameraDeviceDetails {
    camera_device_id: number;
    camera_device_name: string;
}